Seguridad Agregada

+++ ArtSwapv1/src/main/java/pe/edu/upc/artswapv1/repositories/IRolUsuarioRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import pe.edu.upc.artswapv1.entities.Rolusuario;
 
 @Repository
-public interface IRolUsuarioRepository extends JpaRepository<Rolusuario,Integer> {
+public interface IRolUsuarioRepository extends JpaRepository<Rolusuario,Long> {
 
 }
